requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
torch==2.2.2
Pillow==10.2.0
segmentation-models-pytorch==0.3.3
numpy==1.24.3
scikit-learn==1.4.1.post1
joblib==1.3.2
torchvision==0.17.2
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-image==0.23.1